src/image_preprocessor/requirements.txt
envyaml
urllib3==1.26.19
requests
pyOpenSSL
json2html
flask
prometheus_client
schedule
kafka-python-ng
prometheus-client
boto3
pyjwt